backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0

# Validación y serialización (actualizado para compatibilidad)
pydantic>=2.9.2,<3.0.0
email-validator==2.1.0

# Autenticación y seguridad
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Utilidades
python-decouple==3.8
python-multipart==0.0.6

# Dependencias adicionales para FastAPI
httpx>=0.24.0
jinja2>=3.1.0
